pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "otel-logging-framework"
version = "1.0.1"
description = "A comprehensive OpenTelemetry-compliant logging framework for Python"
readme = "README.md"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "filelock>=3.9",
    "elasticsearch==9.0.2",
    "psycopg2-binary>=2.9",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/yourusername/otel-logging-framework"
Repository = "https://github.com/yourusername/otel-logging-framework"
Documentation = "https://github.com/yourusername/otel-logging-framework#readme"

[tool.setuptools]
packages = ["otel_logger", "otel_logger.handlers"]
include-package-data = true

[tool.setuptools.package-data]
otel_logger = ["*.yaml"]